Change device file to 32 bit FIFO, IP does not work as expected

+++ mp4/Mat_mult/matrixmath.h
@@ -5,7 +5,7 @@
 #define HW_COSIM //Allow hardware software co-simulation
 
 #define MATRIX_MAX_SIZE		50
-void MAT_Stream(float *in_arr, float *out_arr, char op_type);
+void MAT_Stream(int *in_arr, int *out_arr, char op_type);
